elementalconductor: set XMLName on Preset structure
This is a bugfix when creating a preset on Elemental
Conductor API. We were sending the preset XML with
<Preset> instead of <preset> (with P capitalized)
because of the lack of a XMLName field on the Preset
structure.

This commit also fixes the return type of createPreset()
call. Now we return the preset created in conformity with
createJob().